requirements.txt
transformers==4.15.0
discord.py==1.7.3
pydantic==1.8.2